pyproject.toml
[project]
name = "bike-shop"
version = "0.1.0"
description = "Bike Shop team agents — multi-tool personas powered by Claude Code CLI"
requires-python = ">=3.12"
dependencies = [
    "slack-bolt>=1.18",
    "slack-sdk>=3.27",
    "python-dotenv>=1.0",
    "PyJWT[crypto]>=2.8",
    "anthropic>=0.40.0",
    "langfuse>=2.0",
    "mem0ai>=0.1",
    "ollama>=0.4",
    "redis==5.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
bike-shop = "bike_shop.main:main"
